pyproject.toml
[project]
name = "rec-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.2",
    "django>=6.0",
    "django-storages[boto3]>=1.14.6",
    "djangorestframework>=3.16.1",
    "redis>=7.1.0",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "factory-boy>=3.3.3",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
    "requests>=2.32.5",
    "ruff>=0.14.9",
    "tqdm>=4.67.1",
]

[tool.pytest.ini_options]
pythonpath = ["src"] 
DJANGO_SETTINGS_MODULE = "recommendation_system.settings"
